refactor: Services to work with shoppingcart in WEBUI
