--- src/MerchStore.WebUI/Models/ShoppingCart/ShoppingCartItem.cs
@@ -1,10 +1,0 @@
-namespace MerchStore.WebUI.Models.ShoppingCart;
-
-public class ShoppingCartItem
-{
-    public Guid ProductId { get; set; }
-    public string Name { get; set; }
-    public decimal PriceAmount { get; set; }
-    public int Quantity { get; set; }
-    public string? ImageUrl { get; set; }
-}
